+++ internal/controller/operator/factory/vlcluster/vlstorage.go
@@ -240,6 +240,9 @@ func buildVLStoragePodSpec(cr *vmv1.VLCluster) (*corev1.PodTemplateSpec, error) 
 
 	vmMounts = append(vmMounts, cr.Spec.VLStorage.VolumeMounts...)
 
+	volumes, vmMounts = build.LicenseVolumeTo(volumes, vmMounts, cr.Spec.License, vmv1beta1.SecretsDir)
+	args = build.LicenseArgsTo(args, cr.Spec.License, vmv1beta1.SecretsDir)
+
 	for _, s := range cr.Spec.VLStorage.Secrets {
 		volumes = append(volumes, corev1.Volume{
 			Name: k8stools.SanitizeVolumeName("secret-" + s),